new file:   app/src/app/auth/layout.tsx 	new file:   app/src/app/auth/login/page.tsx 	new file:   app/src/app/auth/signup/page.tsx 	modified:   app/src/app/layout.tsx 	deleted:    app/src/app/login/page.tsx 	modified:   app/src/app/page.tsx 	deleted:    app/src/app/signup/page.tsx 	modified:   app/src/components/forms/LoginForm.tsx 	modified:   app/src/components/forms/SignupForm.tsx 	modified:   app/src/components/navigation/Navbar.tsx
