+++ app/src/app/auth/layout.tsx
@@ -1,0 +1,11 @@
+export default function AuthLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return (
+    <main className="flex-1 flex justify-center">
+       {children} 
+    </main>
+  );
+}
