+++ app/src/app/auth/login/page.tsx
@@ -1,0 +1,9 @@
+import { Dropdown, TextInput } from "@/components/input";
+import Navbar from "@/components/navigation/Navbar";
+import LoginForm from "@/components/forms/LoginForm";
+
+export default function Login() {
+  return (
+      <LoginForm />
+  );
+}
