+++ app/src/app/auth/signup/page.tsx
@@ -1,0 +1,9 @@
+import { Dropdown, TextInput } from "@/components/input";
+import Navbar from "@/components/navigation/Navbar";
+import SignupForm from "@/components/forms/SignupForm";
+
+export default function Signup() {
+  return (
+      <SignupForm />
+  );
+}
