+++ app/src/app/layout.tsx
@@ -1,5 +1,6 @@
 import "./globals.css";
 import { Roboto, Alegreya } from "next/font/google";
+import Navbar from "@/components/navigation/Navbar";
 
 const inter = Alegreya({ subsets: ["latin"], variable: "--font-alegreya" });
 const roboto = Roboto({
@@ -21,9 +22,17 @@ export default function RootLayout({
   return (
     <html lang="en">
       <body
-        className={`px-4 pb-8 container mx-auto justify-center ${inter.variable} ${roboto.variable}`}
+        className={`flex flex-col min-h-screen px-4 pb-8 container mx-auto justify-center ${inter.variable} ${roboto.variable}`}
       >
+        <Navbar />
         {children}
+        <footer className="lg:px-12 flex justify-between">
+            <div className="flex flex-col gap-2 font-medium capitalize tracking-wider">
+                <div className="  ">PRIVACY & LEGAL</div>
+                <div className="">CONTACT</div>
+            </div>
+            <div>© 2023 - One Day, All rights reserved.</div>
+        </footer>
       </body>
     </html>
   );
